services/ml-service/src/main.py
"""
Service ML pour traitement NLP bilingue (FR/MG)
API REST FastAPI pour l'analyse de sentiment, classification de contenu, etc.
"""

from fastapi import FastAPI, HTTPException, BackgroundTasks
from fastapi.middleware.cors import CORSMiddleware
from pydantic import BaseModel
from typing import List, Optional, Dict, Any
import uvicorn
from loguru import logger
import os
from dotenv import load_dotenv

# Chargement des variables d'environnement
load_dotenv()

# Configuration de l'application
app = FastAPI(
    title="AI4Local ML Service",
    description="Service de Machine Learning pour traitement NLP FR/MG",
    version="1.0.0",
    docs_url="/docs",
    redoc_url="/redoc"
)

# Configuration CORS
app.add_middleware(
    CORSMiddleware,
    allow_origins=["*"],  # À restreindre en production
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

# Modèles Pydantic
class TextAnalysisRequest(BaseModel):
    text: str
    language: str  # 'fr' ou 'mg'
    analysis_type: str  # 'sentiment', 'classification', 'keywords'

class TextAnalysisResponse(BaseModel):
    text: str
    language: str
    analysis_type: str
    result: Dict[str, Any]
    confidence: float
    processing_time: float

class ContentOptimizationRequest(BaseModel):
    content: str
    target_audience: str
    platform: str  # 'facebook', 'instagram', 'email'
    language: str

class ContentOptimizationResponse(BaseModel):
    original_content: str
    optimized_content: str
    suggestions: List[str]
    seo_score: float
    readability_score: float

class HealthResponse(BaseModel):
    status: str
    version: str
    models_loaded: List[str]

# Services (à implémenter)
class MLService:
    def __init__(self):
        self.models = {}
        logger.info("Initialisation du service ML")
    
    async def load_models(self):
        """Chargement des modèles ML au démarrage"""
        try:
            # Charger les modèles pré-entraînés
            logger.info("Chargement des modèles ML...")
            # self.models['sentiment_fr'] = load_sentiment_model('fr')
            # self.models['sentiment_mg'] = load_sentiment_model('mg')
            # self.models['classifier'] = load_classification_model()
            logger.info("Modèles ML chargés avec succès")
        except Exception as e:
            logger.error(f"Erreur lors du chargement des modèles: {e}")
    
    async def analyze_text(self, request: TextAnalysisRequest) -> TextAnalysisResponse:
        """Analyse de texte (sentiment, classification, extraction de mots-clés)"""
        import time
        start_time = time.time()
        
        try:
            result = {}
            confidence = 0.85  # Placeholder
            
            if request.analysis_type == "sentiment":
                # Analyse de sentiment
                result = {
                    "sentiment": "positive",  # positive, negative, neutral
                    "score": 0.85,
                    "emotions": {
                        "joy": 0.6,
                        "confidence": 0.3,
                        "surprise": 0.1
                    }
                }
            elif request.analysis_type == "classification":
                # Classification du contenu
                result = {
                    "category": "commercial",
                    "subcategory": "promotion",
                    "topics": ["produit", "vente", "marketing"]
                }
            elif request.analysis_type == "keywords":
                # Extraction de mots-clés
                result = {
                    "keywords": ["marketing", "digital", "madagascar", "PME"],
                    "entities": {
                        "organizations": ["AI4Local"],
                        "locations": ["Madagascar"],
                        "products": []
                    }
                }
            
            processing_time = time.time() - start_time
            
            return TextAnalysisResponse(
                text=request.text,
                language=request.language,
                analysis_type=request.analysis_type,
                result=result,
                confidence=confidence,
                processing_time=processing_time
            )
            
        except Exception as e:
            logger.error(f"Erreur lors de l'analyse de texte: {e}")
            raise HTTPException(status_code=500, detail="Erreur lors de l'analyse")
    
    async def optimize_content(self, request: ContentOptimizationRequest) -> ContentOptimizationResponse:
        """Optimisation de contenu pour plateformes spécifiques"""
        try:
            # Logique d'optimisation selon la plateforme
            optimized_content = request.content  # Placeholder
            
            if request.platform == "facebook":
                # Optimisation Facebook
                suggestions = [
                    "Ajouter un call-to-action",
                    "Utiliser des hashtags pertinents",
                    "Réduire la longueur du texte"
                ]
            elif request.platform == "instagram":
                # Optimisation Instagram
                suggestions = [
                    "Ajouter plus d'émojis",
                    "Structurer avec des hashtags",
                    "Optimiser pour mobile"
                ]
            else:
                suggestions = ["Contenu optimisé pour email"]
            
            return ContentOptimizationResponse(
                original_content=request.content,
                optimized_content=optimized_content,
                suggestions=suggestions,
                seo_score=75.5,
                readability_score=82.3
            )
            
        except Exception as e:
            logger.error(f"Erreur lors de l'optimisation: {e}")
            raise HTTPException(status_code=500, detail="Erreur lors de l'optimisation")

# Instance du service ML
ml_service = MLService()

# Routes API
@app.on_event("startup")
async def startup_event():
    """Initialisation au démarrage"""
    logger.info("Démarrage du service ML")
    await ml_service.load_models()

@app.get("/health", response_model=HealthResponse)
async def health_check():
    """Vérification de l'état du service"""
    return HealthResponse(
        status="healthy",
        version="1.0.0",
        models_loaded=list(ml_service.models.keys())
    )

@app.post("/api/v1/analyze", response_model=TextAnalysisResponse)
async def analyze_text(request: TextAnalysisRequest):
    """Analyse de texte NLP"""
    logger.info(f"Analyse de texte: {request.analysis_type} - {request.language}")
    return await ml_service.analyze_text(request)

@app.post("/api/v1/optimize", response_model=ContentOptimizationResponse)
async def optimize_content(request: ContentOptimizationRequest):
    """Optimisation de contenu"""
    logger.info(f"Optimisation de contenu pour {request.platform}")
    return await ml_service.optimize_content(request)

@app.get("/api/v1/models")
async def list_models():
    """Liste des modèles disponibles"""
    return {
        "models": list(ml_service.models.keys()),
        "languages_supported": ["fr", "mg"],
        "analysis_types": ["sentiment", "classification", "keywords"]
    }

if __name__ == "__main__":
    port = int(os.getenv("PORT", 5000))
    logger.info(f"🐍 ML Service démarré sur le port {port}")
    uvicorn.run(
        "main:app",
        host="0.0.0.0",
        port=port,
        reload=True if os.getenv("ENV") == "development" else False
    )